refactor: add unmount for savannah
