+++ src/components/savannah/common/savannah.common.constans.js
@@ -1,1 +1,9 @@
 export const DIGIT_CODES = ['Digit1', 'Digit2', 'Digit3', 'Digit4', 'Numpad1', 'Numpad2', 'Numpad3', 'Numpad4'];
+
+export const GAME_DEFAULT = {
+  BurningLivesCount: 5,
+  WordGroupRange: { Min: 0, Max: 5 },
+  WordPageRange: { Min: 0, Max: 9 },
+  WordsPerExampleSentence: 20,
+  WordsPerPage: 60,
+};
